Refactor
Separated code for calculating the double_gyre dynamics into its own file.

Reduced episode limit from 2000 to 200 because of the increased run time of the new design.

Note: this code is still time-independent. The new design is set up for time dependence, but double_gyre is still called with t=0 at all steps. I was just testing that the refactor did not change any of my previous results.

Result Set 5 was generated from this version of the code.